perf(import): confine every no-cycle search to the importer's group
Withholding dynamic edges does not break the correspondence the confinement
rests on. The groups are computed over the whole edge set and the search walks
a subset of it, so every file a route can pass through both reaches the
importer and is reached from it, which is what sharing a group means.
Confining the walk can therefore only prune files no route reaches, whatever
allowUnsafeDynamicCyclicDependency is set to.

That leaves the option paying the same per-file cost as every other
configuration instead of searching the whole reachable graph. On a synthetic
project of one 200-file component whose members each import a 20-deep chain,
the whole-project search goes from 63.5ms to 3.1ms; on 200 small components of
the same total size, 1.35ms to 0.18ms. The configurations that were already
confined are unchanged.

The two predicates the top-level loop applied then say the same thing, and
collapse into the one isReportCandidate answers.

The search moves in with the graph it walks: it is a property of that graph
rather than of the file that started it, and no_cycle.go is left with the rule
— options, entry point, message.
